Moved Facebook, Instagram, Twitter, VK cleanup functions from "generator" to "indexer". Added call to ValidateAndFormat_* functions on social links edit. Fixed contact:line validation. Empty string is now correct value
