+++ indexer/contacts_processor.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <string>
+
+namespace osm {
+  std::string ValidateAndFormat_facebook(std::string const & v);
+  std::string ValidateAndFormat_instagram(std::string const & v);
+  std::string ValidateAndFormat_twitter(std::string const & v);
+  std::string ValidateAndFormat_vk(std::string const & v);
+  std::string ValidateAndFormat_contactLine(std::string const & v);
+}
